+++ CHANGELOG.md
@@ -1,3 +1,13 @@
+## [0.1.1](https://github.com/EndyKaufman/ngx-bind-io/compare/0.1.0...0.1.1) (2019-01-20)
+
+
+### Bug Fixes
+
+* Add prefix for detect used inputs and outputs ([03d8acd](https://github.com/EndyKaufman/ngx-bind-io/commit/03d8acd))
+* Trim used array to usedInputs and usedOutputs ([ebbfb16](https://github.com/EndyKaufman/ngx-bind-io/commit/ebbfb16))
+
+
+
 # 0.1.0 (2019-01-20)